% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{clean_out_plot}
\alias{clean_out_plot}
\title{applies custom modifications to ggplot object}
\usage{
clean_out_plot(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
The modified ggplot object
}
\description{
applies custom modifications to ggplot object
}
