% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_plotting.R
\name{h2o_extract_classification_errors}
\alias{h2o_extract_classification_errors}
\title{extracts classification errors from h2o.scoreHistory()}
\usage{
h2o_extract_classification_errors(mod)
}
\arguments{
\item{mod}{An h2o classification problem}
}
\value{
Classification errors
}
\description{
extracts classification errors from h2o.scoreHistory()
}
