% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{msg_color}
\alias{msg_color}
\title{returns a specific msg color from the palette}
\usage{
msg_color(color = "red")
}
\arguments{
\item{color}{A color name or index}
}
\description{
returns a specific msg color from the palette
}
