% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_plotting.R
\name{plot_h2o_ROC}
\alias{plot_h2o_ROC}
\title{plots ROC from one or multiple fitted h2o models}
\usage{
plot_h2o_ROC(...)
}
\arguments{
\item{...}{one or more trained h2o models}
}
\value{
A ggplot2 object
}
\description{
plots ROC from one or multiple fitted h2o models
}
