% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_apply_msg_colors}
\alias{theme_apply_msg_colors}
\title{applies the msg color palette}
\usage{
theme_apply_msg_colors(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
The modified ggplot object
}
\description{
applies the msg color palette
}
