% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_remove_background}
\alias{theme_remove_background}
\title{set blank background on ggplot object}
\usage{
theme_remove_background(p)
}
\arguments{
\item{p}{A ggplot object}
}
\description{
set blank background on ggplot object
}
