% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_remove_grid}
\alias{theme_remove_grid}
\title{remove all gridlines from ggplot object}
\usage{
theme_remove_grid(p)
}
\arguments{
\item{p}{A ggplot object}
}
\description{
remove all gridlines from ggplot object
}
