% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_rotate_labels_x}
\alias{theme_rotate_labels_x}
\title{rotate x axis labels}
\usage{
theme_rotate_labels_x(p, angle = 90)
}
\arguments{
\item{angle}{The angle by which labels should be rotated. Defaults to 90.}
}
\description{
@inheritParams clean_out_plot
}
