% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_rotate_labels_x_reset}
\alias{theme_rotate_labels_x_reset}
\title{resets axis rotation to 0}
\usage{
theme_rotate_labels_x_reset(p)
}
\arguments{
\item{p}{A ggplot object}
}
\description{
resets axis rotation to 0
}
