% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_show_axes}
\alias{theme_show_axes}
\title{Show axes as solid black lines}
\usage{
theme_show_axes(p)
}
\arguments{
\item{p}{a ggplot object}
}
\value{
The modified ggplot object
}
\description{
Show axes as solid black lines
}
