% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_show_axis_x}
\alias{theme_show_axis_x}
\title{show x axis as solid black line}
\usage{
theme_show_axis_x(p)
}
\arguments{
\item{p}{a ggplot object}
}
\description{
show x axis as solid black line
}
