% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_show_axis_y}
\alias{theme_show_axis_y}
\title{show y axis as solid black line}
\usage{
theme_show_axis_y(p)
}
\arguments{
\item{p}{a ggplot object}
}
\description{
show y axis as solid black line
}
