% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_plotting.R
\name{plot_h2o_classification_error}
\alias{plot_h2o_classification_error}
\title{plots test and validation error extracted from an h2o object}
\usage{
plot_h2o_classification_error(mod)
}
\arguments{
\item{mod}{An h2o.ai model}
}
\description{
plots test and validation error extracted from an h2o object
}
